# CSV parsing internals refactoring