--- a/src/CSV.jl
+++ b/src/CSV.jl
@@ -4,6 +4,12 @@ if !isdefined(Base, :contains)
     contains(haystack, needle) = occursin(needle, haystack)
 end
 
+if applicable(Core.Compiler.typesubtract, Union{Int, Missing}, Missing)
+    ts(T, S) = Core.Compiler.typesubtract(T, S)
+else
+    ts(T, S) = Core.Compiler.typesubtract(T, S, 16)
+end
+
 # stdlib
 using Mmap, Dates, Unicode
 # Parsers.jl is used for core type parsing from byte buffers
@@ -16,6 +22,8 @@ using PooledArrays
 # SentinelArrays.jl allow efficient conversion from Vector{Union{T, Missing}} to Vector{T}
 # it also provides the MissingVector and ChainedVector array types
 using SentinelArrays
+using WeakRefStrings
+export PosLenString
 
 struct Error <: Exception
     msg::String
@@ -23,9 +31,13 @@ end
 
 Base.showerror(io::IO, e::Error) = println(io, e.msg)
 
+# constants
+const SINGLE_THREADED_POOL_DEFAULT = 0.75
+const MULTI_THREADED_POOL_DEFAULT = 0.25
+
 include("utils.jl")
 include("detection.jl")
-include("header.jl")
+include("context.jl")
 include("file.jl")
 include("chunks.jl")
 include("rows.jl")