--- a/src/CSV.jl
+++ b/src/CSV.jl
@@ -4,12 +4,6 @@ if !isdefined(Base, :contains)
     contains(haystack, needle) = occursin(needle, haystack)
 end
 
-if applicable(Core.Compiler.typesubtract, Union{Int, Missing}, Missing)
-    ts(T, S) = Core.Compiler.typesubtract(T, S)
-else
-    ts(T, S) = Core.Compiler.typesubtract(T, S, 16)
-end
-
 # stdlib
 using Mmap, Dates, Unicode
 # Parsers.jl is used for core type parsing from byte buffers
@@ -32,8 +26,8 @@ end
 Base.showerror(io::IO, e::Error) = println(io, e.msg)
 
 # constants
-const SINGLE_THREADED_POOL_DEFAULT = 0.75
-const MULTI_THREADED_POOL_DEFAULT = 0.25
+const DEFAULT_STRINGTYPE = InlineString
+const DEFAULT_POOL = 0.25
 const DEFAULT_LINES_TO_CHECK = 30
 const DEFAULT_MAX_WARNINGS = 100
 const TRUE_STRINGS = ["true", "True", "TRUE", "T", "1"]