--- a/src/chunks.jl
+++ b/src/chunks.jl
@@ -1,10 +1,6 @@
-struct Chunks{H}
-    h::H # Header
-    threaded::Union{Bool, Nothing}
-    typemap::Dict{Type, Type}
-    tasks::Int
-    debug::Bool
-    ranges::Vector{Int64}
+struct Chunks
+    ctx::Context
+    Chunks(ctx::Context) = new(ctx)
 end
 
 """
@@ -37,13 +33,12 @@ function Chunks(source;
     footerskip::Integer=0,
     transpose::Bool=false,
     comment::Union{String, Nothing}=nothing,
-    use_mmap=nothing,
     ignoreemptylines::Bool=true,
-    threaded::Union{Bool, Nothing}=nothing,
-    tasks::Integer=Threads.nthreads(),
-    lines_to_check::Integer=30,
     select=nothing,
     drop=nothing,
+    limit::Union{Integer, Nothing}=nothing,
+    tasks::Integer=Threads.nthreads(),
+    lines_to_check::Integer=DEFAULT_LINES_TO_CHECK,
     # parsing options
     missingstrings=String[],
     missingstring="",
@@ -56,31 +51,43 @@ function Chunks(source;
     dateformat::Union{String, Dates.DateFormat, Nothing}=nothing,
     dateformats::Union{AbstractDict, Nothing}=nothing,
     decimal::Union{UInt8, Char}=UInt8('.'),
-    truestrings::Union{Vector{String}, Nothing}=["true", "True", "TRUE"],
-    falsestrings::Union{Vector{String}, Nothing}=["false", "False", "FALSE"],
+    truestrings::Union{Vector{String}, Nothing}=TRUE_STRINGS,
+    falsestrings::Union{Vector{String}, Nothing}=FALSE_STRINGS,
     # type options
     type=nothing,
     types=nothing,
     typemap::Dict=Dict{Type, Type}(),
-    pool::Union{Bool, Real}=0.1,
+    pool::Union{Bool, Real, AbstractVector, AbstractDict}=NaN,
     lazystrings::Bool=false,
+    stringtype::StringTypes=String,
     strict::Bool=false,
     silencewarnings::Bool=false,
+    maxwarnings::Int=DEFAULT_MAX_WARNINGS,
     debug::Bool=false,
     parsingdebug::Bool=false)
 
-    h = Header(source, header, normalizenames, datarow, skipto, footerskip, transpose, comment, use_mmap, ignoreemptylines, select, drop, missingstrings, missingstring, delim, ignorerepeated, quotechar, openquotechar, closequotechar, escapechar, dateformat, dateformats, decimal, truestrings, falsestrings, type, types, typemap, pool, lazystrings, strict, silencewarnings, debug, parsingdebug, false)
-    rowsguess, ncols, buf, len, datapos, options = h.rowsguess, h.cols, h.buf, h.len, h.datapos, h.options
-    N = tasks > rowsguess || rowsguess < 100 ? 1 : tasks == 1 ? 8 : tasks
-    chunksize = div(len - datapos, N)
-    ranges = Int64[i == 0 ? datapos : i == N ? len : (datapos + chunksize * i) for i = 0:N]
-    findrowstarts!(buf, len, options, ranges, ncols, h.types, h.flags, lines_to_check)
-    return Chunks(h, threaded, typemap, tasks, debug, ranges)
+    ctx = Context(source, header, normalizenames, datarow, skipto, footerskip, transpose, comment, ignoreemptylines, select, drop, limit, true, tasks, lines_to_check, missingstrings, missingstring, delim, ignorerepeated, quotechar, openquotechar, closequotechar, escapechar, dateformat, dateformats, decimal, truestrings, falsestrings, type, types, typemap, pool, lazystrings, stringtype, strict, silencewarnings, maxwarnings, debug, parsingdebug, false)
+    !ctx.threaded && throw(ArgumentError("unable to iterate chunks from input file source"))
+    foreach(col -> col.lock = ReentrantLock(), ctx.columns)
+    return Chunks(ctx)
 end
 
+Base.IteratorSize(::Type{Chunks}) = Base.HasLength()
+Base.length(x::Chunks) = x.ctx.ntasks
+Base.eltype(x::Chunks) = File{false}
+
 function Base.iterate(x::Chunks, i=1)
-    i >= length(x.ranges) && return nothing
-    f = File(x.h; finalizebuffer=false, startingbyteposition=x.ranges[i], endingbyteposition=(x.ranges[i + 1] - (i != length(x.ranges))), threaded=x.threaded, typemap=x.typemap, tasks=x.tasks, debug=x.debug)
+    i > x.ctx.ntasks && return nothing
+    names = copy(x.ctx.names)
+    columns = [Column(col) for col in x.ctx.columns]
+    datapos = x.ctx.chunkpositions[i]
+    len = x.ctx.chunkpositions[i + 1] - 1
+    rowsguess = cld(x.ctx.rowsguess, x.ctx.ntasks)
+    threaded = false
+    ntasks = 1
+    limit = typemax(Int64)
+    ctx = Context(x.ctx.transpose, x.ctx.name, names, rowsguess, x.ctx.cols, x.ctx.buf, datapos, len, 1, x.ctx.options, x.ctx.coloptions, columns, x.ctx.pool, x.ctx.customtypes, x.ctx.typemap, x.ctx.stringtype, limit, threaded, ntasks, x.ctx.chunkpositions, x.ctx.maxwarnings, x.ctx.debug, x.ctx.streaming)
+    f = File(ctx, true)
     return f, i + 1
 end