@bkamins 2021-05-23
So after this change there will be no `PooledString` option. Right? We should use `pool` everywhere.

From the description it was not clear to me what happens if the user chooses `pool=true` for some column but also `lazystrings=true`. What will be the type of the produced column if pooling is done?

Also how is `PosLenStringVector` different from the old `LazyStringVector`? I thought it would be more flexible (i.e. allow subsetting, resizing etc. as long as we are not trying to change the pool of strings used)