@quinnj 2021-05-24
> From the description it was not clear to me what happens if the user chooses `pool=true` for some column but also `lazystrings=true`. What will be the type of the produced column if pooling is done?

`lazystrings=true` is now deprecated, and just sets `stringtype=PosLenString` internally. So type of pooled column will match whatever is passed as `stringtype=X` (to allow for short strings, `String`, `PosLenString`, etc.).

> Also how is PosLenStringVector different from the old LazyStringVector? I thought it would be more flexible (i.e. allow subsetting, resizing etc. as long as we are not trying to change the pool of strings used)

`PosLenStringVector` is very similar to `LazyStringVector`, but will produce `PosLenString` elements on indexing, rather than `String`, thus avoiding full `String` allocations when using. It also supports a few limited operations via `similar`, `copyto!`, `getindex` w/ multiple inds that will produce `PosLenStringVector` instead of producing full `Vector{PosLenString}`.