@bkamins 2021-06-01
On:
```
(DF_DEV) pkg> st
      Status `C:\WORK\dev\DF_DEV\Project.toml`
  [69666777] Arrow v1.5.0
  [6e4b80f9] BenchmarkTools v1.0.0
  [336ed68f] CSV v0.8.5 `https://github.com/JuliaData/CSV.jl.git#jq/internalsrefactor`
  [324d7699] CategoricalArrays v0.10.0
  [a93c6f00] DataFrames v1.1.1 `C:\Users\bogum\.julia\dev\DataFrames`
  [2dfb63ee] PooledArrays v1.2.1 `https://github.com/JuliaData/PooledArrays.jl.git#main`
  [ea10d353] WeakRefStrings v0.6.2 `https://github.com/JuliaData/WeakRefStrings.jl.git#jq/inlinestrings`
```
I do
```
df = DataFrame([[randstring(i) for _ in 1:1000] for i in 1:20], :auto)
CSV.write("test.csv", df)
```
and then I get on Windows:
```
julia> df2 = CSV.read("test.csv", DataFrame)
ERROR: MethodError: no method matching (PooledArrays.PooledArray{WeakRefStrings.InlineString1, R, N, RA} where {R<:Integer, N, RA})(::PooledArrays.RefArray{Vector{UInt32}}, ::Dict{Union{Missing, WeakRefStrings.InlineString1}, UInt32})
Stacktrace:
 [1] makepooled!(col::CSV.Column)
```
Which seems to signal that you are using some branch of PooledArrays.jl. Which one should be used?