@bkamins 2021-06-01
Ah PooledArrays.jl#bkamins-patch-1 is required. Now it works.