@bkamins 2021-06-01
Below I report on the case where we have a regression.

The benchmarks in https://github.com/JuliaData/Arrow.jl/issues/188 for `N=10^7` are (I have used `pool=false` for now as the current state of how we use PooledArrays.jl should be changed):

this set of PRs:
```
julia> GC.gc(false); @time innerjoin(x, small, on = :id1, makeunique=true, matchmissing=:equal);
  0.402371 seconds (328 allocations: 849.167 MiB)

julia> GC.gc(false); @time leftjoin(x, medium, on = :id2, makeunique=true, matchmissing=:equal);
  0.567584 seconds (433 allocations: 1.095 GiB, 2.45% gc time)

julia> GC.gc(false); @time innerjoin(x, big, on = :id3, makeunique=true, matchmissing=:equal);
  3.729259 seconds (371 allocations: 1.237 GiB, 0.45% gc time)
```

current released versions:
```
julia> GC.gc(false); @time innerjoin(x, small, on = :id1, makeunique=true, matchmissing=:equal);
  0.417128 seconds (328 allocations: 780.502 MiB)

julia> GC.gc(false); @time leftjoin(x, medium, on = :id2, makeunique=true, matchmissing=:equal);
  0.547001 seconds (433 allocations: 1.002 GiB)

julia> GC.gc(false); @time innerjoin(x, big, on = :id3, makeunique=true, matchmissing=:equal);
  3.306644 seconds (371 allocations: 1.103 GiB)
```

These tests mostly check copying of `InlineString`. As you can see we have a small regression in `innerjoin` on `big`. The reason is (at least this is what I think) that we have there:
1. one column with `InlineString15`
2. two columns with `InlineString7`

If we are not making operations on these strings then these columns are more expensive to copy than just `Vector{String}` because they have a larger memory footprint (instead of pointers we store data which is larger).

This means that in the examples above (where the data is small enough not to trigger GC) using `String` is faster.