@bkamins 2021-06-05
I think that given the amount of changes introduced it is essential not only to update docstrings, but also to prepare a list of examples of common scenarios the user might want to use (e.g. I am currently unsure what we would have pass in H2O benchmarks to make sure all numeric columns are loaded normally and all string columns are pooled independent on the count of unique values and loaded as inline strings).