@quinnj 2021-06-06
> to make sure all numeric columns are loaded normally and all string columns are pooled independent on the count of unique values and loaded as inline strings

Numeric columns will not be pooled automatically. The only way to pool a numeric column is to request it specifically like `pool=Dict(1 => true)`. Passing `pool=true` means all string, and only string, columns _will_ be pooled, regardless of cardinality. Currently, the default is `stringtype=InlineString`, which means columns with individual string sizes < 32 will be inline string types, while columns with larger string values will be `String`.

So all that to say: you shouldn't have to pass anything by default. Unless you really want to make sure all string columns are pooled, but the logic should be much better now to automatically choose pooling or not. But probably the only thing we need to pass is `CSV.read(file, DataFrame; pool=true)`.

But I agree that I want to have a much better "manual" kind of documentation that walks through the various configuration options.