@bkamins 2021-06-06
Thank you! I have forgotten to add that we should also read in integers as `Int32`, the example current command is:
```
CSV.read(src_grp, DataFrame, types=[PooledString, PooledString, PooledString, Int32, Int32, Int32, Int32, Int32, Float64], threaded=false);
```