@quinnj 2021-06-06
> I have forgotten to add that we should also read in integers as `Int32`

Ah, interesting. Is that just for memory savings? Or does it make processing faster too? It wouldn't be too hard to allow reading `Int32` columns by default and promote to Int64 if needed.