@bkamins 2021-06-06
> Ah, interesting. Is that just for memory savings? Or does it make processing faster too?

Both - it uses less memory and is cheaper to copy (and both are relevant in the benchmarks).

> It wouldn't be too hard to allow reading `Int32` columns by default and promote to `Int64` if needed.

I think it is OK to read in `Int64` as default, I just need to know how in the new API we should pass information that we want integers to be read in as `Int32`.