@dkaszynski 2021-06-10
> I think it is OK to read in `Int64` as default, I just need to know how in the new API we should pass information that we want integers to be read in as `Int32`.

I agree that `Int64` as a default make sense. However, it would be great if there was an option similar to Pandas `downcast`, where for integer columns you can get the smallest type that can hold the data in the given column. I often find it very useful when working with data that has very large volumes.