@quinnj 2021-06-15
Ok, pushed code to support reading Int32 by default and promoting to Int64 if necessary.