@bkamins 2021-06-15
> Ok, pushed code to support reading Int32 by default and promoting to Int64 if necessary.

I think it is also OK to use this approach, as `Int32` should be enough in most cases as a default.

@nalimilan - agreed?